server/requirements.txt
Flask==1.1.1
pymongo==3.9.0
python-dateutil==2.8.0
